refactor(storage): Remove ktx imports
